+++ java/brusselnieuws-rss/brusselnieuws-rss-datastore-model/src/test/java/be/seriousbusiness/brusselnieuws/rss/datastore/model/dto/impl/AuthorDTOImplTest.java
@@ -1,11 +1,13 @@
 package be.seriousbusiness.brusselnieuws.rss.datastore.model.dto.impl;
 
+import java.math.BigInteger;
+
 import be.seriousbusiness.brusselnieuws.rss.datastore.model.dto.AbstractAuthorDTOTest;
 
 public class AuthorDTOImplTest extends AbstractAuthorDTOTest<AuthorDTOImpl> {
 
 	public AuthorDTOImplTest() {
-		super(121121121l,"© brusselnieuws.be");
+		super(BigInteger.valueOf(121121121l),"© brusselnieuws.be");
 	}
 
 	@Override
