+++ java/brusselnieuws-rss/brusselnieuws-rss-datastore-model/src/test/java/be/seriousbusiness/brusselnieuws/rss/datastore/model/dto/impl/CategoryDTOImplTest.java
@@ -1,11 +1,13 @@
 package be.seriousbusiness.brusselnieuws.rss.datastore.model.dto.impl;
 
+import java.math.BigInteger;
+
 import be.seriousbusiness.brusselnieuws.rss.datastore.model.dto.AbstractCategoryDTOTest;
 
 public class CategoryDTOImplTest extends AbstractCategoryDTOTest<CategoryDTOImpl> {
 
 	public CategoryDTOImplTest() {
-		super(321321321l,"http://www.brusselnieuws.be/nl/media-type/artikel","artikel");
+		super(BigInteger.valueOf(321321321l),"http://www.brusselnieuws.be/nl/media-type/artikel","artikel");
 	}
 
 	@Override
