+++ java/brusselnieuws-rss/brusselnieuws-rss-datastore-model/src/test/java/be/seriousbusiness/brusselnieuws/rss/datastore/model/dto/stub/MediumTypeDTOStub.java
@@ -1,16 +1,18 @@
 package be.seriousbusiness.brusselnieuws.rss.datastore.model.dto.stub;
 
+import java.math.BigInteger;
+
 import be.seriousbusiness.brusselnieuws.rss.datastore.model.dto.MediumTypeDTO;
 
 public class MediumTypeDTOStub implements MediumTypeDTO {
 
 	@Override
-	public Long getId() {
-		return 123123123l;
+	public BigInteger getId() {
+		return BigInteger.valueOf(123123123l);
 	}
 
 	@Override
-	public void setId(final Long id) {}
+	public void setId(final BigInteger id) {}
 
 	@Override
 	public String getType() {
